requirements.txt
python-hcl2
PyYAML
rich
